+++ src/service/EmergencyDutyService.js
@@ -1,4 +1,5 @@
-import EmergencyDuty from '../domain/EmergencyDuty.js';
+// import EmergencyDuty from '../domain/EmergencyDuty.js';
+import EmergencyDutyScheduler from '../domain/EmergencyDutyScheduler.js';
 import Nicknames from '../domain/Nicknames.js';
 import Schedule from '../domain/Schedule.js';
 
@@ -24,7 +25,7 @@ class EmergencyDutyService {
   }
 
   emergencyDuty() {
-    return new EmergencyDuty(this.#schedule, this.#weekdayNicknames, this.#weekendNicknames).emergencyDuty();
+    return new EmergencyDutyScheduler(this.#schedule, this.#weekdayNicknames, this.#weekendNicknames).emergencyDuty();
   }
 }
